requirements.txt
anyio==4.11.0
certifi==2025.10.5
charset-normalizer==3.4.4
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.11
python-dotenv==1.1.1
python-telegram-bot==21.6
requests==2.32.3
fuzzywuzzy==0.18.0
python-levenshtein==0.27.1
sniffio==1.3.1
typing_extensions==4.15.0
urllib3==2.5.0
flask==3.0.0
